+++ src/AppendStream.php
@@ -176,6 +176,10 @@ final class AppendStream implements StreamInterface
      */
     public function read(int $length): string
     {
+        if ($this->streams === []) {
+            return '';
+        }
+
         $buffer = '';
         $total = count($this->streams) - 1;
         $remaining = $length;